Implementation of a Main Menu (Day 20(?))
Hello

This will be pushed in a new branch in my repo for now for which I will be implementing a GUI component in my project.

The main goal of this project is to be a full fledged app, and that wouldn't be possible if it is only an application in the console. So, moving forward I'm going to learn Java Swing and implement a GUI for this project of mine.

That is all.